+++ sdk/compute/azure-resourcemanager-compute-bulkactions/src/main/java/com/azure/resourcemanager/compute/bulkactions/implementation/StartResourceOperationResponseImpl.java
@@ -4,13 +4,11 @@
 
 package com.azure.resourcemanager.compute.bulkactions.implementation;
 
-import com.azure.resourcemanager.compute.bulkactions.fluent.models.ResourceOperationInner;
 import com.azure.resourcemanager.compute.bulkactions.fluent.models.StartResourceOperationResponseInner;
 import com.azure.resourcemanager.compute.bulkactions.models.ResourceOperation;
 import com.azure.resourcemanager.compute.bulkactions.models.StartResourceOperationResponse;
 import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
 
 public final class StartResourceOperationResponseImpl implements StartResourceOperationResponse {
     private StartResourceOperationResponseInner innerObject;
@@ -36,11 +34,9 @@ public final class StartResourceOperationResponseImpl implements StartResourceOp
     }
 
     public List<ResourceOperation> results() {
-        List<ResourceOperationInner> inner = this.innerModel().results();
+        List<ResourceOperation> inner = this.innerModel().results();
         if (inner != null) {
-            return Collections.unmodifiableList(inner.stream()
-                .map(inner1 -> new ResourceOperationImpl(inner1, this.manager()))
-                .collect(Collectors.toList()));
+            return Collections.unmodifiableList(inner);
         } else {
             return Collections.emptyList();
         }