+++ sdk/identity/azure-identity/CHANGELOG.md
@@ -20,6 +20,15 @@
 
 - Upgraded `msal4j` from `1.23.1` to `1.26.0`.
 
+## 1.18.6 (2026-08-31)
+
+### Other Changes
+
+#### Dependency Updates
+
+- Upgraded `azure-core` from `1.59.0` to version `1.59.1`.
+- Upgraded `azure-core-http-netty` from `1.16.6` to version `1.16.7`.
+
 ## 1.18.5 (2026-08-24)
 
 ### Other Changes